tests: Extract test_message_topics.py for topic tests.
